requirements.txt
# ---------- Core app ----------
streamlit==1.37.1
pandas==2.2.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# ---------- TF 2.15 stack (no keras here; we'll add it separately) ----------
tensorflow==2.15.0
tensorflow-intel==2.15.0
tensorflow-estimator==2.15.0
tensorboard==2.15.2
tensorflow-io-gcs-filesystem==0.31.0
# NumPy range compatible with TF 2.15 + our DICOM libs:
numpy>=1.23.5,<2.0
protobuf==4.25.8
h5py==3.15.1
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
libclang==18.1.1
ml-dtypes==0.2.0
opt-einsum==3.4.0
optree==0.17.0
six==1.17.0
termcolor==3.1.0
werkzeug==3.1.3
markdown==3.9
tensorboard-data-server==0.7.2

# ---------- DICOM ----------
pydicom==2.4.4
# pin pylibjpeg family to NumPy 1.x-compatible versions:
pylibjpeg==2.0.1
pylibjpeg-openjpeg==2.2.0
pylibjpeg-libjpeg==2.2.0
# (keep gdcm out unless you really need it; it often fails on Windows)

# ---------- PyTorch (CPU) + XRV ----------
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchxrayvision==1.2.4
scikit-image==0.24.0
imageio==2.37.0
scipy==1.15.3
sympy==1.14.0
tqdm==4.67.1
networkx==3.3
filelock==3.19.1
fsspec==2025.9.0
jinja2==3.1.6

# ---------- Streamlit deps pinned ----------
pyarrow==22.0.0
rich==13.9.4
tornado==6.5.2
toml==0.10.2
typing-extensions==4.15.0
watchdog==4.0.2
altair==5.5.0
blinker==1.9.0
cachetools==5.5.2
click==8.3.0
gitpython==3.1.45
pydeck==0.9.1
requests==2.32.5
tenacity==8.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.28.0
pygments==2.19.2
narwhals==2.9.0
idna==3.11
charset-normalizer==3.4.4
urllib3==2.5.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
rsa==4.9.1
smmap==5.0.2
colorama==0.4.6
packaging==24.2
mdurl==0.1.2
markdown-it-py==4.0.0
lazy-loader==0.4
tbb==2021.11.0
